Moves get log columns to item log view component
